web/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.6
redis==5.0.1
huggingface-hub==0.20.3
sqlalchemy==2.0.27
pydantic==2.6.1
pyyaml==6.0.1
python-dotenv==1.0.1